app/requirements.txt
python-dotenv
django
djangorestframework
django-filter
PyJWT
Pillow
